code formatting and cleanup
